automatic_database_backup/requirements.txt
Flask==2.3.3
Flask-Admin==1.6.1
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Mail==0.9.1
Flask-CORS==4.0.0
gspread==5.11.3
google-auth==2.22.0
google-api-python-client==2.108.0
google-cloud-storage==2.10.0
gunicorn==21.2.0
apscheduler==3.10.4
python-dotenv==1.0.0
psycopg2-binary==2.9.7
alembic==1.12.1
pandas==2.0.3
openpyxl==3.1.2
pytz==2023.3
